Now the ping works in both ways, also added hash map to store the sent queries inside node, to map them to appropriate replies

+++ Kademlia/include/node_entry.h
@@ -10,7 +10,7 @@
 
 namespace kademlia {
     struct NodeEntry {
-        uint16_t port_;
+        int port_;
         QHostAddress ip_address_;
         std::bitset<constants::kNodeIdSize> node_id_;
